Refactor[blockstack]: Replace fakeAuth with blockstackAuth
