+++ src/index.js
@@ -1,98 +1,12 @@
 // Dependencies
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {
-  BrowserRouter as Router,
-  Route,
-  Link,
-  Redirect,
-  withRouter
-} from 'react-router-dom';
-
-// Styles
-import './index.css';
-
-// Private App Routes
-import PrivateRoutes from './private-routes'
-
 import registerServiceWorker from './registerServiceWorker';
-
-const fakeAuth = {
-  isAuthenticated: false,
-  authenticate(cb) {
-    this.isAuthenticated = true
-    setTimeout(cb, 100) // fake async
-  },
-  signout(cb) {
-    this.isAuthenticated = false
-    setTimeout(cb, 100) // fake async
-  }
-}
-
-class Login extends React.Component {
-  state = {
-    redirectToReferrer: false
-  }
-
-  login = () => {
-    fakeAuth.authenticate(() => {
-      this.setState(() => ({
-        redirectToReferrer: true
-      }))
-    })
-  }
-
-  render() {
-    const { from } = this.props.location.state || { from: { pathname: '/p/profile' } }
-    const { redirectToReferrer } = this.state
-
-    if (redirectToReferrer === true) {
-      return <Redirect to={from} />
-    }
-
-    return (
-      <div>
-        <p>You must log in to view the page</p>
-        <button onClick={this.login}>Log in</button>
-      </div>
-    )
-  }
-}
-
-const PrivateRoute = ({ component: Component, ...rest }) => (
-  <Route {...rest} render={(props) => (
-    fakeAuth.isAuthenticated === true
-      ? <Component {...props} />
-      : <Redirect to={{
-          pathname: '/login',
-          state: { from: props.location }
-        }} />
-  )} />
-)
-
-const AuthButton = withRouter(({ history }) => (
-  fakeAuth.isAuthenticated ? (
-    <p>
-      Welcome! <button onClick={() => {
-        fakeAuth.signout(() => history.push('/'))
-      }}>Sign out</button>
-    </p>
-  ) : (
-    <p>You are not logged in.</p>
-  )
-))
+// Blockstack app
+import App from './App'
 
 ReactDOM.render(
-  <Router>
-    <div>
-      <AuthButton/>
-      <ul>
-        <li><Link to="/p">Protected Page</Link></li>
-      </ul>
-      <Route path="/login" component={Login}/>
-      <PrivateRoute path='/p' component={PrivateRoutes} />
-    </div>
-  </Router>,
+  <App />,
   document.getElementById('root'));
 
 registerServiceWorker();
